requirements.txt
addict
future
lmdb
matplotlib
numpy
opencv-python
Pillow
pyyaml
requests
scikit-image
scipy
tqdm
yapf
sklearn
pandas
gpustat
tensorboard
seaborn
wandb
